+++ ignite/services/scaffolder/scaffolder.go
@@ -69,10 +69,13 @@ func finish(ctx context.Context, cacheStorage cache.Storage, path, gomodPath str
 		return err
 	}
 
-	if err := gocmd.ModTidy(ctx, path); err != nil {
+	if err := gocmd.Fmt(ctx, path); err != nil {
 		return err
 	}
-	return gocmd.Fmt(ctx, path)
+
+	_ = gocmd.GoImports(ctx, path) // goimports installation could fail, so ignore the error
+
+	return gocmd.ModTidy(ctx, path)
 }
 
 func protoc(ctx context.Context, cacheStorage cache.Storage, projectPath, gomodPath string) error {